update analytics chart modules
